refactor: use target_compile_options instead of global CMAKE_CXX_FLAGS*
For now we want to always compile without optimizations because
we found that using `-O2` causes issues selecting data from tables.
Uur mariadb.spec adds `-O2` to CMAKE_CXX_FLAGS_RELWITHDEBINFO
when the build type is RELWITHDEBINFO (which is the default).
This then causes this storage engine to build with -O2
optimizations becuase CMAKE_CXX_FLAGS_RELWITHDEBINFO are
added to the end of CMAKE_CXX_FLAGS.

We can work around this issue by appending -O0 to
CMAKE_CXX_FLAGS_RELWITHDEBINFO which will then cause -O0
to override the -O2 set previously but this pollutes the
global CMAKE_CXX_FLAGS_RELWITHDEBINFO variable and causes everything
to be built unoptimized. Instead we can use target_compile_options
which only pass compile options for this target.